oreder, product  수정

+++ uni-place-backend/src/main/java/org/myweb/uniplace/domain/commerce/repository/OrderItemRepository.java
@@ -3,5 +3,5 @@ package org.myweb.uniplace.domain.commerce.repository;
 import org.myweb.uniplace.domain.commerce.domain.entity.OrderItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderItemRepository extends JpaRepository<OrderItem, Long> {
+public interface OrderItemRepository extends JpaRepository<OrderItem, Integer> {  // ✅ Fix: Long → Integer
 }
